OrdersDAO, Order Model and new logic
Introduced a OrdersDAO which is a data access object for accessing the orders in the Shopify database. The class currently provides two methods, one for fetching the data object containing orders meeting the criteria, and another method, that gets the data object, but then maps it to the Orders model and returns an array of Order objects, for simplicity and convenience.,
Also introduced the Order object, which represents an Order item fetched from the Shopify Database, it only provides what is needed.