New structure for the Orders Object and Product object
Changed the structure of the Orders model class, so each order is encapsulating all of the products, that belong to it in a an array. This way I make sure each order is responsible for carring information about each product it consists of.
Changed the structure of the Product model class too. The model now encapsulates product weight, variation id and product quantity. This provides a more structured way of handling the products, the variations and their quantities