CollectionsDAO and Collections Map
Moved the functions taking care of the collections, formerly in the ShopifyClient class, now in a separate DAO class, responsible for taking care of the Collections object on the Shopify database.
Moved further with the logic of calculating the products in kilograms sold for each DPA collection. Wrote a BLOC that is in charge of getting all the collections products and setting them to a map, together with which collection id they belong to. This implements the process of having a map (storage) where each collection knows which product IDs belong to it.