Re-factored how the app handles weight
As the primary goal of the application is to calculate how many kilograms of each product category we have sold, I re-factored the code, so it pushes to the Products sold map the kilograms that have been sold.
The weight in grams is being stored in a temporary object inside the method, this time I don't iterate over the map more than once and therefore achieve a better prefomance of the algorithm cutting down the time it takes to calculate the weights and push them to the map.