Changed the calculation logic and split a method
Changed the logic for calculating the products total weigh in the ProducsSoldMap class. Instead of calculating the products weight in kilogram within the initialize method, I moved the logic to the CollectionsTotalWeightMap class. This way, I avoid any errors that might be caused on the long way of several calculations. As javascript has problems with calculating floating points, I found it a more suitable solution to calculcate the floating point (kilograms) as last calculation, before returning the result.
Also split the logic of the CollectionsTotalWeightMap class' method for getting the total weight of the collections into two smaller methods. I introduced a method that calculates the weight and then another mmethod converts it into kilograms and returns it. The logic could be split further into one more smaller method