Set up two unit test to test ProductsSoldMap and CollectionsTotalWeightMap
Set up two very basic tests to test the functionality of the two main classes calculating the weight of all products for each collection. The tests were set up with Jest testing library. Both tests mock data and avoid contacting external sources of information. The tests are located in the test folder in the root of the project.
The ProductsSoldMap test is testing whether the total weight for each products is calculated correctly. The test mocks orders data.
The CollectionsTotalWeightMap test test whether the total weight for each collection is calculated properly, based on mock Orders and Products. The