Added a new method for getting orders above the shopify limit
To get all orders for a period, exceeding the retrieval limit set by shopify, I needed to write a method that implements pagination logic with the since_id parameter provided on the method by the official Shopify-api library.  The method utilizes a while loop, to loop and call the method for orders retreival untill there are no more orders received, which is logically the end of the orders list in the Shopify database. The method gets the last order's id and passes it as a since_id parameter on the next call.