Added a Remix app as a presentation layer and set up an endpoint
Installed express to set up express server and an endpoint for accessing the API from the client side. Installed CORS to make sure the cross origin requests are allowed and possible. Currently the server has only one endpoint and it's for getting an array of collection weights.
Set up a Remix app through the Shopify CLI. Wrote a service for accessing the DPA calculator server. Currently the Remix app does not have anything.