Changed the architecture a little and added CollectionsAgent
Changed the architecture (folder structure) of the ServiceLayer of the application.
Also introduced a new service, which is the CollectionsAgent, Currently the agent has only one method, that is for creating a test collection in the Shopify store. The method is using the graphQL API of Shopify, as the Rest one does not provide endpoints for creating collections. The agent is to be extended further.