Changed teh CreateCollections module and updated other modules
Changed the CreateCollections module to be CollectionsManager, all interactions with the collections Shopify object will happen through the CollectionsManager service.  The CollectionsManager class currently has two methods createCollectionsFor and getWeeeCollections. The createCollectionsFor method takes and argument a Map and creates a collection for each of the pairs in the map in the vendor's store and the getWeeeCollections method takes in an array of collection names and returns an array with their ids
Moved the createCollection method to the CollectionsGraphDAO module.