Changed the logic and the structure of the CollectionsMap module
Instanciated all of the methods from the CollectionsDAO in the CollectionsManager, this way, there is a single entry point for all methods related to the Collections Shopify object.
Instead of hardcoding the values of the collection ids that are to be summed up, I implemented the logic for getting them differently, by utilising the method for getting collection id's based on name from the CollectionsManager getWeeeCollections():
There is currently an error that is being thrown by the initilize() method part of the CollectionsMap class, that is to be fixed later!