Added a method to add products to a collection in CollectionsManager
Needed a method for adding products to a collection. First created a method in the CollectionsGraphDAO as adding products to collections is possible through the Shopify GraphQL api. The method takes in collectionId, which is the id of the collection that will get the products and an array of product ids.