Modified CollectionsTotalWeightMap class
Added two new methods to the CollectionsTotalWeightMap class. Separated the logic for converting the weights from grams to kilograms, instead of the logic for converting the units living within the lifitime of the getCollectionsTotalWeight() method, it was enacpsulated in a separate method convertCollectionsUnit().
Intoduced a new method for converting the collections's ids to collection titles. The method is utilising the collectionsManager' getCollectionNameById().