Started re-writing and re-factoring the architectural structure of the application
As previously the application was cluttered with tangled code and tightly coupled classes which significantly troubles the re-factoring and scaling of the code base.
The application changes structure, mostly because the authentication token needs to be sent with every request and not saved as it is currently.  The whole class tree changes, as the authentication token needs to be passed from a class to super class, all the way to the Shopify Client.