Introduced Utilities and a CollectionsCalculator service classs
Introduced Utilities  directory and RequestUtilities class, which is in charge of encapsulating different helper methods, to be used throughout the API methods. This way I ensure less code clutter and simplified main.ts file, without unnecessary logic in the API route functions. Currently the RequestUtilities  class has only one method, which is for extracting headers from the request object.
Introduced a CollectionsCalculator service, which is in charge of callculating the collections total weight and returning a map with it. By  intorducing this class, I reduce business logic in the API route functions.
Changed the directory struycture a little, in the sake of readability.