Started setting up tests, to test the new architecture
Started fixing the old, already implemented tests, to be up to date with the new architecture of the system. The tests are still not performing, as they are not fully re-implemented.
Introduced a test, to test the CollectionCalculator service. The test currently has some non-finished mocks. To be finished later