Routes re-factoring custom error class
Improved the different API endpoints part of the main.ts. Included more and better status code responses for different cases.
Wrote a ResourceNotFoundException, which is a custom error class, to throw exceptions, when a resource wasn't found in the Shopify Database. This way I ensure better errors and smoother debugging experience.
Added CollectionsProductService, which currently contains a method for adding products to a collection, based on collection name