Re-factored app and added new route
Re-factored the DAOs that use the graphQL api of shopify's to use the new request method, instead of the legacy method "query".
Added a new method in the CollectionsGraphDAO for getting all collections out of the vendor's store. The method is called getAllCollections
Added a method, for getting just the WEEE collections in the CollectionsManager class. The method is called getWeeeCollections and it utilises the getAllCollections of the CollectionsGraphDAO. The method filters the collections and returnes only the ones that contain "WEEE" in their name.
Added an api endpoint for getting the WEEE collections from vendor's store.