Modification to include report fromDate and toDate.
Modified the CollectionsTotalWeightMap class, the CollectionsMap, the ProductsSoldMap and the OrdersDAO to expect from and to date, in the methods resposnible for initialising the calculation of the report. In the OrdersDAO, modified the getOrdersBetween, which previosly had the valued for fromDate and toDate hardcoded to require them as a parameters passed to the method.
In the CollectionsTotalWeightMap modified the calculateCollectionsTotalWeight(), initialize() and getCollectionsTotalWeight() method to include fromDate and toDate as parameters.
Additinoally, cleaned up the code space and renamed some of the class members with more descriptive names.