Refactored the server, to now include OrdersManager class in the service layer
Introduced an OrdersGraphDAO class, that handles interactions with the Orders object through the Shopify GraphQL API. Currently the class contains only one method getOrdersCountFor(), which returns the orders count for a venor's store based on specified time frame and target country. Needed to also modify the DaoFactory, to also include the OrdersGraphDAO.
Added an OrdersManager class to handle all interactions with the Orders object. The class contains all needed methods from both the OrdersDAO and the OrdersGraphDAO class