Fixed logic flow fault in main and added better responses in class methods
Fixed a logic (control flow) fault in main within the /initCalculation route. An if statement was incorrectly set to execute the route logic, when accessToken and hostName are missing.
Started refactoring the responses methods of DAOs and Managers return. Some methods didn't have a specific return type. Methods are now getting a return type of {isSuccess: boolean, payload. error?: string}. Where isSuccess notifies the higher level logic, that calls the method, that the action is a success, error is optional and might not exist on all methods and payload is the payload data returned from the method.