Continued modifying response types in methods
Modified mmethods in the Collecitons DAOs and in the CollectionsManager.

Additionally refactored the "initialize" and "updateCollectionsMap" in the CollectionsMap method, to also check for success and to throw errors, instead of simply returning null;

Modified the control flow logic of some routes in main, to utilise better the new response types and also I made sure they implement proper response status codes to the consumer.