Minor changes in logic of how different methods handle responses
Changed the  calculateCollectionsTotalWeight method in the CollectionsCalculator service class, to return a better response type, including isSuccess property and error.
Additionally, modified the  getCollectionsTotalWeight method in CollectionsTotalWeightMap class to also match the response type. Now it also includes isSuccess and error properties.
Implemented a proper error handling on the /initCalculation route in the main module.