Fixed the issue with the /addProductsToCollection route
To fix the issue with the /addProductsToCollection route modified the  addProductsToCollection method part of the CXollectionsProductService, to not get collection id by name, as the collection id is passed to the server on request from client side.
Additionally modified the query in addProductsToCollection method within the DAO to not create smart collections, but manual collections instead, as the smart collection does not allow manual assignment of products. Within the same method removed the composition of id, as id is passed in a proper and usable form.