Finished the setup of the health routes
Finished the sett up of the health routes.
Now both health routes should log errors or general logs.
The database health route checkls if the server is connected to PGsql properly. The health check is intended only for production, will not work locally, as the server is not connected to any database locally. The database health check returns an object with timestamp and database status.
The server health check route checks the health of the server itself. The route also checks database health, however the database health does not affect the status of the server health check. The route sends an object to the client, containing information about status, timestamp, uptime and a dependencies object (where the database status is returned)