Logic for shop/redact webhook.
Wrote the logic for deleting stored shop record from the database - finished the deleteShopRecord method in the PgSqlAccessor class. The method deletes the record from the database with a sql query. Deletes the record WHERE the shop field equals the host argument passed to the method.
Modified slightly the verifyShopifyWebhook method from WebhookUtils file. The method checks if rawBody has been sent, if not, throws an error.
Changed the way the webhook routes are verifying. The verifyShopifyWebhook method is passed as  a first method in the routes. The method will return a status code, if  verification failed.
Additionally changed the way the dbConnectionObject is composed in PGConnection in the initConnection() method. Now the PGSQL_PASSWORD is parsed as string when passed to the object.